requirements.txt
requests
flask
python-telegram-bot